Organize perf tasks
