+++ cmds/modules/noded/main.go
@@ -21,6 +21,8 @@ import (
 	"github.com/threefoldtech/zos/pkg/events"
 	"github.com/threefoldtech/zos/pkg/monitord"
 	"github.com/threefoldtech/zos/pkg/perf"
+	"github.com/threefoldtech/zos/pkg/perf/cpubench"
+	"github.com/threefoldtech/zos/pkg/perf/iperf"
 	"github.com/threefoldtech/zos/pkg/registrar"
 	"github.com/threefoldtech/zos/pkg/stubs"
 	"github.com/threefoldtech/zos/pkg/utils"
@@ -202,10 +204,10 @@ func action(cli *cli.Context) error {
 		return errors.Wrap(err, "failed to create a new perfMon")
 	}
 
-	iperfTest := perf.NewIperfTest()
+	iperfTest := iperf.NewIperfTest()
 	perfMon.AddTask(&iperfTest)
 
-	cpuBenchmarkTask := perf.NewCPUBenchmarkTask()
+	cpuBenchmarkTask := cpubench.NewCPUBenchmarkTask()
 	perfMon.AddTask(&cpuBenchmarkTask)
 
 	if err = perfMon.Run(ctx); err != nil {
