+++ pkg/perf/iperf/graphql_nodes.go
@@ -1,4 +1,4 @@
-package perf
+package iperf
 
 import (
 	"context"
