+++ pkg/perf/iperf/iperf_task.go
@@ -1,4 +1,4 @@
-package perf
+package iperf
 
 import (
 	"context"
